requirements.txt
# 这是用conda自动导出的环境要求，助教测试时可以有需自取
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.1.1
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
cryptography=44.0.2=pypi_0
expat=2.6.4=h8ddb27b_0
fastapi=0.115.12=pypi_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
idna=3.10=pypi_0
iniconfig=2.1.0=pypi_0
libffi=3.4.4=hd77b12b_1
openssl=3.0.16=h3f729d1_0
packaging=24.2=pypi_0
pip=25.0=py312haa95532_0
pluggy=1.5.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pyjwt=2.10.1=pypi_0
pytest=8.3.5=pypi_0
python=3.12.9=h14ffc60_0
requests=2.32.3=pypi_0
setuptools=75.8.0=py312haa95532_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.46.1=pypi_0
tk=8.6.14=h0416ee5_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025a=h04d1e81_0
urllib3=2.4.0=pypi_0
uvicorn=0.34.0=pypi_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wheel=0.45.1=py312haa95532_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
